Dive into events in JavaScript.md
Dive into events in JavaScript
In addition to the emergence of events, there is also an immersion. In fact, the event first goes from top to bottom (intercept stage), reaches our element (target stage) and only then starts to float (bubbling stage).

In order to hang an event handler taking into account the stage of interception, addEventListener has a third optional parameter. If it is equal to true - the event will fire at the stage of interception, and if false - at the stage of bubbling (this is by default). Let's look at an example: